Added changes to create common file for util functions
