+++ pkg/inframanager/test/services_test.go
@@ -17,9 +17,7 @@
 package test
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
 	"flag"
 	"fmt"
 	p4_v1 "github.com/p4lang/p4runtime/go/p4/v1"
@@ -82,61 +80,6 @@ var (
 var (
 	serviceIpAddr = "10.10.100.1"
 )
-
-type UUIDGenerator struct {
-	idGen        uint32
-	internalChan chan uint32
-}
-
-func newUUIDGenerator() *UUIDGenerator {
-	gen := &UUIDGenerator{
-		idGen:        0,
-		internalChan: make(chan uint32, DEFAULT_UUID_CNT_CACHE),
-	}
-	gen.startGen()
-	return gen
-}
-
-// Open goroutine and put the generated UUID in digital form into the buffer pipe
-func (this *UUIDGenerator) startGen() {
-	go func() {
-		for {
-			if this.idGen == MAXUINT32 {
-				this.idGen = 1
-			} else {
-				this.idGen += 1
-			}
-			this.internalChan <- this.idGen
-		}
-	}()
-}
-
-// Get UUID in uint32 form
-func (this *UUIDGenerator) getUUID() uint32 {
-	return <-this.internalChan
-}
-
-var uuidFactory = newUUIDGenerator()
-
-func valueToBytes(value uint32) []byte {
-	buf := new(bytes.Buffer)
-	err := binary.Write(buf, binary.BigEndian, uint32(value))
-	if err != nil {
-		fmt.Println("binary.Write failed:", err)
-	}
-	fmt.Printf("% x", buf.Bytes())
-	return buf.Bytes()
-}
-
-func valueToBytes16(value uint16) []byte {
-	buf := new(bytes.Buffer)
-	err := binary.Write(buf, binary.BigEndian, value)
-	if err != nil {
-		fmt.Println("binary.Write failed:", err)
-	}
-	fmt.Printf("% x", buf.Bytes())
-	return buf.Bytes()
-}
 
 func MacToPortTable(ctx context.Context, p4RtC *client.Client, macAddr string, port uint32, flag bool) error {
 	var err error
